seperate into 2 api
