+++ api/src/main/java/eu/darkbot/api/managers/AgatusAPI.java
@@ -1,0 +1,8 @@
+package eu.darkbot.api.managers;
+
+/**
+ * API for any Agatus event
+ */
+
+public interface AgatusAPI extends DefaultNpcEventAPI {
+}
